requirements.txt
PySide6>=6.6.0
mss>=9.0.1
Pillow>=10.0.0
pynput>=1.7.6
openai>=1.10.0
google-generativeai>=0.6.0  # <-- 新增
keyring>=23.5.0
pydantic>=2.5.0
qasync>=0.25.0
aiohttp>=3.9.1
# Platform specific dependencies for accurate window capture
pywin32; platform_system=="Windows"